Revert "merge OpenRoad?master with naveen"

+++ src/odb/src/db/tmg_conn.cpp
@@ -62,8 +62,11 @@ tmg_conn::tmg_conn(utl::Logger* logger) : logger_(logger)
   _csVV.reserve(1024);
   _csNV.reserve(1024);
   _shortV.reserve(1024);
-  _need_short_wire_id = false;
+  _cut_end_extMin = 1;
+  _need_short_wire_id = 0;
   _first_for_clear = nullptr;
+  _preserveSWire = false;
+  _swireNetCnt = 0;
 }
 
 tmg_conn::~tmg_conn() = default;
@@ -75,7 +78,11 @@ int tmg_conn::ptDist(const int fr, const int to) const
 
 tmg_rcpt* tmg_conn::allocPt(int x, int y, dbTechLayer* layer)
 {
-  return &_ptV.emplace_back(x, y, layer);
+  tmg_rcpt* pt = &_ptV.emplace_back();
+  pt->_x = x;
+  pt->_y = y;
+  pt->_layer = layer;
+  return pt;
 }
 
 void tmg_conn::addRc(const dbShape& s,
@@ -173,7 +180,9 @@ void tmg_conn::addITerm(dbITerm* iterm)
   _csVV.emplace_back();
   _csNV.emplace_back();
 
-  tmg_rcterm& x = _termV.emplace_back(iterm);
+  tmg_rcterm& x = _termV.emplace_back();
+  x._iterm = iterm;
+  x._bterm = nullptr;
   x._pt = nullptr;
   x._first_pt = nullptr;
 }
@@ -183,14 +192,19 @@ void tmg_conn::addBTerm(dbBTerm* bterm)
   _csVV.emplace_back();
   _csNV.emplace_back();
 
-  tmg_rcterm& x = _termV.emplace_back(bterm);
+  tmg_rcterm& x = _termV.emplace_back();
+  x._iterm = nullptr;
+  x._bterm = bterm;
   x._pt = nullptr;
   x._first_pt = nullptr;
 }
 
 void tmg_conn::addShort(const int i0, const int i1)
 {
-  _shortV.emplace_back(i0, i1);
+  tmg_rcshort& x = _shortV.emplace_back();
+  x._i0 = i0;
+  x._i1 = i1;
+  x._skip = false;
   if (_ptV[i0]._fre) {
     _ptV[i0]._fre = false;
   } else {
@@ -1322,7 +1336,13 @@ void tmg_conn::analyzeNet(dbNet* net)
     findConnections();
     bool noConvert = false;
     if (_hasSWire) {
-      net->destroySWires();
+      if (_preserveSWire) {
+        net->setDoNotTouch(true);
+        noConvert = true;
+        _swireNetCnt++;
+      } else {
+        net->destroySWires();
+      }
     }
     relocateShorts();
     treeReorder(noConvert);
@@ -1420,7 +1440,7 @@ bool tmg_conn::checkConnected()
 void tmg_conn::treeReorder(const bool no_convert)
 {
   _connected = true;
-  _need_short_wire_id = false;
+  _need_short_wire_id = 0;
   if (_ptV.empty()) {
     return;
   }
@@ -1639,7 +1659,7 @@ void tmg_conn::addToWire(const int fr,
       return;
     }
     if (_ptV[fr]._dbwire_id < 0) {
-      _need_short_wire_id = true;
+      ++_need_short_wire_id;
       return;
     }
     _ptV[to]._dbwire_id = _ptV[fr]._dbwire_id;
@@ -1771,7 +1791,7 @@ void tmg_conn::addToWire(const int fr,
 
   if (_need_short_wire_id) {
     copyWireIdToVisitedShorts(fr);
-    _need_short_wire_id = false;
+    _need_short_wire_id = 0;
   }
 
   int to_id = -1;
